# Improved CreatureVisuals performance by removing viewports.