--- a/project/src/main/viewport-outline-material.tres
+++ b/project/src/main/viewport-outline-material.tres
@@ -6,5 +6,5 @@
 resource_local_to_scene = true
 shader = ExtResource( 1 )
 shader_param/width = 2.5
-shader_param/black = Color( 0.423529, 0.262745, 0.192157, 1 )
+shader_param/black = Color( 0.254902, 0.156863, 0.117647, 1 )
 shader_param/edge_fix_factor = 0.001