--- a/project/src/main/world/creature/Body.tscn
+++ b/project/src/main/world/creature/Body.tscn
@@ -1,0 +1,6 @@
+[gd_scene load_steps=2 format=2]
+
+[ext_resource path="res://src/main/world/creature/body.gd" type="Script" id=3]
+
+[node name="Body" type="Node2D"]
+script = ExtResource( 3 )