… performance-metrics.ods
… project/assets/main/creatures/secondary/terpion.json
… project/assets/main/world/creature/1/neck-blend-packed.json
… project/assets/main/world/creature/1/neck-blend-packed.png
… project/assets/main/world/creature/1/neck-blend-packed.png.import
… project/src/main/editor/creature/CreatureEditor.tscn
… project/src/main/world/creature/Body2Shadows.tscn
… project/src/main/world/creature/Body2Shape.tscn
… project/src/main/world/creature/Creature.tscn
… project/src/main/world/creature/CreatureVisuals.tscn
… project/src/main/world/creature/FatSpriteMover1.tscn
… project/src/main/world/creature/FatSpriteMover2.tscn
… project/src/main/world/creature/MovementPlayer.tscn
… project/src/main/world/creature/belly-colors.gd
… project/src/main/world/creature/body-shadows.shader
… project/src/main/world/creature/body.gd
… project/src/main/world/creature/creature-body.gd
… project/src/main/world/creature/creature-curve.gd
… project/src/main/world/creature/creature-loader.gd
… project/src/main/world/creature/dna-loader.gd
… project/src/main/world/creature/mouth-player.gd
… project/src/main/world/restaurant/RestaurantScene.tscn